add eol_agnostic flag to textwrap.dedent()
This flag enables textwrap.dedent to treat CRLF line endings the same as LF line endings. If the flag is False and the input text contains a line with only 0-n whitespace characters and a CRLF line ending, the dedent function will treat this line as if it is non-empty leading to unexpected behavior. If the flag is True, the dedent funciton will treat the same line the same way it treats the same line except with LF line endings.